+++ src/wtforms/fields/choices.py
@@ -1,4 +1,6 @@
-import itertools
+import warnings
+from dataclasses import dataclass
+from typing import Optional
 
 from wtforms import widgets
 from wtforms.fields.core import Field
@@ -6,9 +8,54 @@ from wtforms.validators import ValidationError
 
 __all__ = (
     "SelectField",
+    "Choice",
     "SelectMultipleField",
     "RadioField",
 )
+
+
+@dataclass
+class Choice:
+    """
+    A dataclass that represents available choices for
+    :class:`RadioField`, :class:`SelectField` and :class:`SelectMultipleField`
+
+    :param value:
+        The value that will be send in the request.
+    :param label:
+        The label of the option.
+    :param render_kw:
+        A dict containing HTML attributes that will be rendered
+        with the option.
+    :param optgroup:
+        The `<optgroup>` HTML tag in which the option will be
+        rendered.
+    """
+
+    value: str
+    label: Optional[str] = None
+    render_kw: Optional[dict] = None
+    optgroup: Optional[str] = None
+    _selected: bool = False
+
+    @classmethod
+    def from_input(cls, input, optgroup=None):
+        if isinstance(input, Choice):
+            if optgroup:
+                input.optgroup = optgroup
+            return input
+
+        if isinstance(input, str):
+            return Choice(value=input, optgroup=optgroup)
+
+        if isinstance(input, tuple):
+            warnings.warn(
+                "Passing SelectField choices as tuples is deprecated and will be "
+                "removed in wtforms 3.3. Please use Choice instead.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+            return Choice(*input, optgroup=optgroup)
 
 
 class SelectFieldBase(Field):
@@ -30,14 +77,8 @@ class SelectFieldBase(Field):
     def iter_choices(self):
         """
         Provides data for choice widget rendering. Must return a sequence or
-        iterable of (value, label, selected) tuples.
+        iterable of Choice.
         """
-        raise NotImplementedError()
-
-    def has_groups(self):
-        return False
-
-    def iter_groups(self):
         raise NotImplementedError()
 
     def __iter__(self):
@@ -49,15 +90,40 @@ class SelectFieldBase(Field):
             _form=None,
             _meta=self.meta,
         )
-        for i, (value, label, checked) in enumerate(self.iter_choices()):
-            opt = self._Option(label=label, id="%s-%d" % (self.id, i), **opts)
-            opt.process(None, value)
-            opt.checked = checked
+        for i, choice in enumerate(self.iter_choices()):
+            opt = self._Option(
+                id="%s-%d" % (self.id, i), label=choice.label or choice.value, **opts
+            )
+            opt.choice = choice
+            opt.checked = choice._selected
+            opt.process(None, choice.value)
             yield opt
 
-    class _Option(Field):
-        checked = False
+    def choices_from_input(self, choices):
+        if callable(choices):
+            choices = choices()
 
+        if choices is None:
+            return None
+
+        if isinstance(choices, dict):
+            warnings.warn(
+                "Passing SelectField choices in a dict deprecated and will be removed "
+                "in wtforms 3.3. Please pass a list of Choice objects with a "
+                "custom optgroup attribute instead.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+
+            return [
+                Choice.from_input(input, optgroup)
+                for optgroup, inputs in choices.items()
+                for input in inputs
+            ]
+
+        return [Choice.from_input(input) for input in choices]
+
+    class _Option(Field):
         def _value(self):
             return str(self.data)
 
@@ -76,44 +142,14 @@ class SelectField(SelectFieldBase):
     ):
         super().__init__(label, validators, **kwargs)
         self.coerce = coerce
-        if callable(choices):
-            choices = choices()
-        if choices is not None:
-            self.choices = choices if isinstance(choices, dict) else list(choices)
-        else:
-            self.choices = None
+        self.choices = self.choices_from_input(choices)
         self.validate_choice = validate_choice
 
     def iter_choices(self):
-        if not self.choices:
-            choices = []
-        elif isinstance(self.choices, dict):
-            choices = list(itertools.chain.from_iterable(self.choices.values()))
-        else:
-            choices = self.choices
-
-        return self._choices_generator(choices)
-
-    def has_groups(self):
-        return isinstance(self.choices, dict)
-
-    def iter_groups(self):
-        if isinstance(self.choices, dict):
-            for label, choices in self.choices.items():
-                yield (label, self._choices_generator(choices))
-
-    def _choices_generator(self, choices):
-        if not choices:
-            _choices = []
-
-        elif isinstance(choices[0], (list, tuple)):
-            _choices = choices
-
-        else:
-            _choices = zip(choices, choices)
-
-        for value, label in _choices:
-            yield (value, label, self.coerce(value) == self.data)
+        choices = self.choices_from_input(self.choices) or []
+        for choice in choices:
+            choice._selected = self.coerce(choice.value) == self.data
+        return choices
 
     def process_data(self, value):
         try:
@@ -138,10 +174,7 @@ class SelectField(SelectFieldBase):
         if not self.validate_choice:
             return
 
-        for _, _, match in self.iter_choices():
-            if match:
-                break
-        else:
+        if not any(choice._selected for choice in self.iter_choices()):
             raise ValidationError(self.gettext("Not a valid choice."))
 
 
@@ -154,18 +187,12 @@ class SelectMultipleField(SelectField):
 
     widget = widgets.Select(multiple=True)
 
-    def _choices_generator(self, choices):
-        if choices:
-            if isinstance(choices[0], (list, tuple)):
-                _choices = choices
-            else:
-                _choices = zip(choices, choices)
-        else:
-            _choices = []
-
-        for value, label in _choices:
-            selected = self.data is not None and self.coerce(value) in self.data
-            yield (value, label, selected)
+    def iter_choices(self):
+        choices = self.choices_from_input(self.choices) or []
+        if self.data:
+            for choice in choices:
+                choice._selected = self.coerce(choice.value) in self.data
+        return choices
 
     def process_data(self, value):
         try:
@@ -190,7 +217,7 @@ class SelectMultipleField(SelectField):
         if not self.validate_choice or not self.data:
             return
 
-        acceptable = {c[0] for c in self.iter_choices()}
+        acceptable = {self.coerce(choice.value) for choice in self.iter_choices()}
         if any(d not in acceptable for d in self.data):
             unacceptable = [str(d) for d in set(self.data) - acceptable]
             raise ValidationError(