+++ CHANGES.rst
@@ -19,6 +19,10 @@ Unreleased
 -   Added shorter format to :class:`~fields.WeekField`
     defaults :pr:`765`
 -   Move to pyproject.toml :pr:`796`
+-   :class:`~fields.SelectField` refactor. Choices tuples and dicts are
+    deprecated in favor of :class:`~fields.Choice` :pr:`739`
+-   ``<option>`` HTML attributes can be passed using
+    :class:`~fields.Choice` :issue:`692` :pr:`739`
 
 Version 3.0.1
 -------------